Split session asset routes
Move annotations and overlays route registration plus their disk persistence out of the main server while preserving the studio overlay flows that reuse the same writers.
